requirements.txt
requests
beautifulsoup4
playwright
flask
openai
PyYAML